requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arm_pytorch_utilities==0.4.3
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==24.10.0
bleach==6.0.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.88
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
d4rl==1.1
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.26
docker-pycreds==0.4.0
etils==1.5.2
executing==1.2.0
Farama-Notifications==0.0.4
fasteners==0.18
fastjsonschema==2.16.3
filelock==3.16.1
flax==0.8.5
fonttools==4.38.0
fqdn==1.5.1
fsspec==2024.12.0
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.30
glfw==2.5.6
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.8.0
humanize==4.11.0
icecream==2.1.3
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-vim==0.15.1
jupyterlab_server==2.19.0
kiwisolver==1.4.4
labmaze==1.0.6
llvmlite==0.43.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml_collections==0.1.1
ml_dtypes==0.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.4
mujoco==3.2.6
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.60.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.2.4
orbax==0.1.1
orbax-checkpoint==0.6.4
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pathtools==0.1.2
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.5
pycparser==2.21
Pygments==2.14.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-seed==0.2.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
scipy==1.10.0
Send2Trash==1.8.0
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.13.1
tensorflow-probability==0.19.0
tensorstore==0.1.69
termcolor==2.2.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.2.0
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
y-py==0.5.9
ypy-websocket==0.8.2
zarr==2.14.2
zipp==3.13.0
